fix: demote log level to debug when local cache insert fails (#6908)
demotes the log level to debug for when local cache insert fails. This
can happen when multiple queries on same stream for same range are
fired, and after getting file meta from pg, all of them try to insert in
local cache, and after one of them succeeds others will fail with sqlite
error of unique constraint violation. This will simply log it on debug
level instead of error, because even if insert fails in local cache, not
a big problem.

---------

Signed-off-by: Yashodhan Joshi <yjdoc2@gmail.com>
Co-authored-by: Yashodhan Joshi <yjdoc2@gmail.com>
Co-authored-by: Hengfei Yang <hengfei.yang@gmail.com>