fix: scheduled pipeline should catch up when there is failures for a timerange (#7350)
For a scheduled pipeline, we have a logic to catch up the delay whenever
there is a delay. For example, say, at 10:00am, there are some errors
while triggering scheduled pipeline, it will retry 3 times and then
retry again after the given frequency (say 2 mins). Now when the error
is resolved at 11:02am, the scheduled pipeline will again try to run the
timerange of 10:00am - 10:02am first. But since there is a bug in the
current code, this 1 hour delay will always be there. At, 11:04am it
will run for 10:02-10:04, at 11:06, 10:04-10:06am etc.

This pr fixes this bug. So, after the delay due to failures, the next
time it runs successfully, it will be able to catch up the timerange by
immediately retriggering the pipelines until it is in sync with the
current time.

---------

Co-authored-by: Taiming Liu <liutaiming3@gmail.com>