feat: support path based value extraction in cipher functions (#7482)
This adds support for path based extraction of value in cipher function.
Now the fn signature has changed to
```
fn(column,key,path)
```

Previous behavior can be achieved using the call
```
fn(column, key, '.')
```

The path must be a string similar to jq paths, but for operating on
array elements, a `*` must be used. For example, for data like
```
{
a : [
        {b : value},
        {b : value},
        {b : value},
    ]
}

```

path will be 
```
a.*.b
```
If given path is not found, or the value on given path is not an array
or array of strings, it will return the column as-is.

---------

Signed-off-by: Yashodhan Joshi <yjdoc2@gmail.com>
Co-authored-by: Yashodhan Joshi <yjdoc2@gmail.com>