fix: don't broadcast process_stream updates to super cluster (#7777)
In super cluster setup, we now send `compact/delete` events with value
`OK` across regions so that other regions can also cleanup the data they
have. One minor bug here is that, when we call `process_stream` (to let
other nodes know that this compactor node is processing this delete
job), it updates the `compact/delete` key with the node id as value.
This also triggers a broadcast to other regions, and other regions hence
store this new change (even though the node id does not exist in other
regions). Hence, there is a possibility that in the other regions, the
same delete job is picked up again by another node.