fix: Query Recommendations - Use Scheduled Retry + Use ingester_service::ingest (#8920)
# PR Type
Fix

# Description
Query recommendations were adjusted to be based on triggers with retries
upon failure. Instead of relying on retry, we are simply skipping this
run for the next generation of query_recommendations.

Triggers and retries were used in hopes that the non-ingester node will
eventually give up and an ingester node may pick up this trigger. This
was a bad idea from the get go. We use ingester_service::ingest now
which finds the ingester node to upload to it regardless of where the
ingestions request comes from.