fix: Alert and Pipelines history recovery + Pagination fix + View fixes (#8935)
# PR type
fix/enchancement

# Description
The pagination issue with alerts and pipeline history is fixed - now
works as normally expected. A few elements of the views were improved
with better choices of representation. A previous PR about styling
changes has removed the Pipelines History related changes partially.
This was recovered and restored to functional state while also taking in
the styling changes.

---------

Co-authored-by: Bhargav <BJP232004@GMAIL.COM>