fix: Fetching enrichment table data gives db error in mysql (#8961)
Currently `get_enrichment_data_from_db` function fails in mysql, because
the struct used for parsing the enrichment table record uses `i64` for
`created_at` field. In mysql, the field is `bigint unsigned`. Hence, the
conversion from `bigint unsigned` to `i64` (`bigint`) fails and as a
result when the enrichment table data is in db, the search query fails
to fetch enrichment table data from the db.