fix: put a limit on the keep alive timeout for scheduled jobs (#9092)
Currently in the alert manager, we run a keep alive thread for every
scheduled job pulled, so that if the timeout for the job is near and the
job is still in progress, we can increase the timeout a bit so that the
job can complete. But in edge cases, if the job itself is complete, but
the keep-alive thread is still not stopped, it will keep running the
keep alive thread forever.

Hence, a limit is added to the keep alive thread, so that it stops after
the timeout.