ci: auto label and milestone (#9384)
See working run here
https://github.com/openobserve/openobserve/actions/runs/19761555623/job/56624451400

Also check the timeline items that mention gh actions bot added the
label and milestone to it.

This adds a CI which will automatically label every PR on approve and
add it to the lowest open milestone (ignoring backlog)
So if current open milestones are backlog, v0.20.0, v0.20.1 , it will
auto add this on approval to v0.20.0 and add `Needs-Testing` label.

---------

Signed-off-by: Yashodhan Joshi <yjdoc2@gmail.com>
Co-authored-by: Yashodhan Joshi <yjdoc2@gmail.com>